create mapper function
